requirements.txt
tensorflow
Flask
numpy
keras
pillow
gunicorn
